Most basic updates working
The method can parse simple parameters such as the boolean status and techtype, and updates the correct ticket based on the given ticket id.

However, updating the associated members and tags is going to be a little more tricky because they are objects. Commented those parts out as they are not working right now.

Will likely need to split some of this work into helper functions to make the unit test less annoying to write.